\name{EvoRAG-internal}

\alias{sisterContinuous_logSpace}
\alias{find.mle.sister}
\alias{MScorrection_MSwithin}

\title{internal EvoRAG functions}

\description{\code{EvoRAG-internal} functions are not typically called by the user}

\details{
These are internal \pkg{EvoRAG} functions, not intended to be called directly by the user. The include the followsing two functions: sisterContinuous_logSpace and find.mle.sister
}
