\name{EvoRAG-internal}

\alias{sisterContinuous_logSpace}
\alias{find.mle.sister}
\alias{MScorrection_MSwithin}
\alias{simulation.analysis}
\alias{sisterContinuous_logSpace_profile_CI}

\title{internal EvoRAG functions}

\description{\code{EvoRAG-internal} functions are not typically called by the user}

\details{
These are internal \pkg{EvoRAG} functions, not intended to be called directly by the user. They include the following functions: sisterContinuous_logSpace, find.mle.sister, and simulation.analysis.
}
